Move users_have_locations to analytics.py
