+++ corehq/apps/locations/tests/test_dbaccessors.py
@@ -2,10 +2,10 @@ from django.test import TestCase
 from corehq.apps.commtrack.tests.util import bootstrap_location_types
 from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.users.models import CommCareUser, WebUser
+from ..analytics import users_have_locations
 from ..dbaccessors import (get_users_by_location_id, get_user_ids_by_location,
                            get_one_user_at_location, get_user_docs_by_location,
-                           get_all_users_by_location, users_have_locations,
-                           get_users_assigned_to_locations)
+                           get_all_users_by_location, get_users_assigned_to_locations)
 from .util import make_loc, delete_all_locations
