+++ corehq/apps/users/views/mobile/users.py
@@ -37,7 +37,7 @@ from corehq.apps.es.queries import search_string_query
 from corehq.apps.hqwebapp.async_handler import AsyncHandlerMixin
 from corehq.apps.hqwebapp.utils import get_bulk_upload_form
 from corehq.apps.locations.models import Location
-from corehq.apps.locations.dbaccessors import users_have_locations
+from corehq.apps.locations.analytics import users_have_locations
 from corehq.apps.users.util import can_add_extra_mobile_workers
 from corehq.apps.custom_data_fields import CustomDataEditor
 from corehq.const import USER_DATE_FORMAT
